+++ src/test/java/com/transferwise/openbanking/client/oauth/RestOAuthClientTest.java
@@ -5,6 +5,7 @@ import com.transferwise.openbanking.client.configuration.AspspDetails;
 import com.transferwise.openbanking.client.error.ApiCallException;
 import com.transferwise.openbanking.client.oauth.domain.AccessTokenResponse;
 import com.transferwise.openbanking.client.oauth.domain.GetAccessTokenRequest;
+import com.transferwise.openbanking.client.test.factory.AspspDetailsFactory;
 import org.hamcrest.CoreMatchers;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
@@ -32,7 +33,7 @@ import org.springframework.web.client.RestTemplate;
 import java.util.stream.Stream;
 
 import static com.transferwise.openbanking.client.test.factory.AccessTokenResponseFactory.aAccessTokenResponse;
-import static com.transferwise.openbanking.client.test.factory.AspspDetailsFactory.aAspspDetails;
+import static com.transferwise.openbanking.client.test.factory.AspspDetailsFactory.aTestAspspDetails;
 
 @ExtendWith(MockitoExtension.class)
 class RestOAuthClientTest {
@@ -62,7 +63,7 @@ class RestOAuthClientTest {
     @Test
     void getAccessToken() throws Exception {
         GetAccessTokenRequest getAccessTokenRequest = GetAccessTokenRequest.clientCredentialsRequest("payments");
-        AspspDetails aspspDetails = aAspspDetails();
+        AspspDetails aspspDetails = AspspDetailsFactory.aTestAspspDetails();
 
         MultiValueMap<String, String> expectedBody = new LinkedMultiValueMap<>();
         getAccessTokenRequest.getRequestBody().forEach(expectedBody::add);
@@ -91,7 +92,7 @@ class RestOAuthClientTest {
     @Test
     void getAccessTokenThrowsApiCallExceptionOnApiCallFailure() {
         GetAccessTokenRequest getAccessTokenRequest = GetAccessTokenRequest.clientCredentialsRequest("payments");
-        AspspDetails aspspDetails = aAspspDetails();
+        AspspDetails aspspDetails = AspspDetailsFactory.aTestAspspDetails();
 
         mockAspspServer.expect(MockRestRequestMatchers.requestTo(aspspDetails.getTokenUrl()))
             .andExpect(MockRestRequestMatchers.method(HttpMethod.POST))
@@ -109,7 +110,7 @@ class RestOAuthClientTest {
         throws Exception {
 
         GetAccessTokenRequest getAccessTokenRequest = GetAccessTokenRequest.clientCredentialsRequest("payments");
-        AspspDetails aspspDetails = aAspspDetails();
+        AspspDetails aspspDetails = AspspDetailsFactory.aTestAspspDetails();
 
         String jsonResponse = objectMapper.writeValueAsString(response);
